Add vanilla FTP workout
Add a vanilla FTP improvement workout that is basically out of the
book. It is per video recommendation of Dylan Johnson.

The LAP marks in the erg file are hand crafted to have laps spanning
the interval WITH recovery, as I usually display section times for the
other stuff (section to go is enough that way).